Configured for MHI-2007 - added mhi-2007 course id - removed nested replies - removed unnecessary csv columns - formatted responses to remove html tags
